basic room and hilt integration

+++ app/src/main/java/de/erikspall/mensaapp/data/sources/remote/api/model/MenuApiModel.kt
@@ -1,5 +1,5 @@
 package de.erikspall.mensaapp.data.sources.remote.api.model
 
 data class MenuApiModel (
-    val id: Int? = null
+    val id: Long? = null
 )
